requirements.txt
fastapi
uvicorn
flask
gunicorn
requests
openai
python-decouple
